requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
yt-dlp>=2023.11.0
openai-whisper>=20231117
googletrans>=4.0.0rc1
elevenlabs>=0.2.26
pydub>=0.25.1
ffmpeg-python>=0.2.0
python-dotenv>=1.0.0
requests>=2.31.0
aiofiles>=23.2.1 